requirements.txt
ipython
black
loguru~=0.5.3
python-dotenv
sqlalchemy~=2.0.25
fastapi~=0.103.0
starlette~=0.27.0
pydantic~=2.5.3
passlib~=1.7.4
pillow
slowapi~=0.1.9
aioredis~=1.3.1
psycopg2-binary
rich~=13.3.5
redis
uvicorn
fastapi-cache2
python-multipart
python-jose